requirements.txt
tensorflow>=2.12
numpy>=1.24
python-chess
tqdm
pyyaml
